Implement a piecemeal tree file system. (#568)
* Implement a piecemeal tree file system.

This allows smaller trees to be integrated into one bigger one.  Good
for distributed tree creation or users who wish to keep a customization
that won't get squashed by an update.

* Exclude tree files because some are too big.

* Add an attribute to help build the order of nodes in the composite tree.

The moddeditemtree.xml file contains the reference tree that I used to
make the individual trees.
The initializer only attempts to extract the tree files if
minecraft.tree is missing and the whole system is only used if InvTweaks
config folder exists.
TreeBuilder updated to read and use treeOrder attribute to fine-tune the
insertion of nodes in the composite tree.

* Make the treeOrder merge smarter when dealing with missing treeOrders.

* Remove modded tree files.

Make the modded tree button go to a repository of modded trees.
Move the merged tree into the system temp folder to minimize the chance
the user would think they could modify it.
Add a readme to the tree folder.
Move all config files to the InvTweaks config folder.
Make a trees folder in the InvTweaks config folder for the tree files.